+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
 ### Fixes
 
 - Fix an issue that document windows sometimes did not shift the initial position from the last window. 
+- Fix an issue that the scrollable area of the editor in vertical layout orientation could be clipped wrongly when scaled.